requirements.txt

Django==5.1.6
django-allauth==65.7.0
django-axes==7.0.2
djangorestframework==3.16.0
psycopg2-binary
python-dotenv
python-decouple
djangorestframework-simplejwt
openai